server/python-simple-server/pyproject.toml
# pyproject.toml - Modern Python Packaging Configuration
# AI-Form Server
# -------------------------------
# PEP 517/518 compliant build configuration
# -------------------------------

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

# -------------------------------
# Project Metadata (PEP 621)
# -------------------------------
[project]
name = "ai-form-server"
version = "1.0.0"
description = "AI-powered form data generation backend server"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "cilang", email = "cilanguser@gmail.com"}
]
maintainers = [
    {name = "cilang", email = "cilanguser@gmail.com"}
]
keywords = [
    "ai",
    "chat",
    "form",
    "data-generation",
    "flask",
    "openai"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed"
]

dependencies = [
    # Web Framework
    "flask>=2.3.0",
    # CORS Support
    "flask-cors>=4.0.0",
    # Rate Limiting
    "flask-limiter>=3.5.0",
    # OpenAI SDK (for AI chat)
    "openai>=1.0.0",
    # HTTP Client
    "httpx>=0.25.0",
    # Retry Logic
    "tenacity>=8.2.0",
    # Environment Variables
    "python-dotenv>=1.0.0",
    # JWT Support
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-timeout>=2.1.0",
    "pytest-xdist>=3.3.0",
    # Test Utilities
    "responses>=0.23.0",
    "freezegun>=1.2.0",
    # Code Quality
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    # Type stubs
    "types-Flask>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/cilang/ai-form"
Documentation = "https://github.com/cilang/ai-form#readme"
Repository = "https://github.com/cilang/ai-form.git"
Issues = "https://github.com/cilang/ai-form/issues"

[project.scripts]
ai-form-server = "ai_form_server.__main__:main"

# -------------------------------
# Setuptools Configuration
# -------------------------------
[tool.setuptools]
package-dir = {"" = "src"}
packages = ["ai_form_server", "ai_form_server.auth", "ai_form_server.services", "ai_form_server.routes"]

[tool.setuptools.package-data]
ai_form_server = ["py.typed"]

# -------------------------------
# Black Configuration (Code Formatter)
# -------------------------------
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

# -------------------------------
# Ruff Configuration (Linter)
# -------------------------------
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise without from inside except
    "ARG001", # unused function argument (common in Flask routes)
]

[tool.ruff.lint.isort]
known-first-party = ["ai_form_server"]

# -------------------------------
# MyPy Configuration (Type Checker)
# -------------------------------
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "flask.*",
    "flask_cors.*",
    "flask_limiter.*",
    "openai.*",
    "httpx.*",
    "tenacity.*",
    "jwt.*",
]
ignore_missing_imports = true

# -------------------------------
# Pytest Configuration
# -------------------------------
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "-ra",
    "--showlocals",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "auth: Authentication tests",
    "security: Security validation tests",
    "performance: Performance tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "unit: Unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
log_cli = true
log_cli_level = "INFO"

# -------------------------------
# Coverage Configuration
# -------------------------------
[tool.coverage.run]
source = ["src/ai_form_server"]
branch = true
omit = [
    "tests/*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
show_missing = true
skip_covered = true

[tool.coverage.html]
directory = "htmlcov"